removes submodule pipeline/ta3ta2-api
The submodule is hosted on GitLab which is not supported for SSH access from CircleCI.  Using HTTPS access for the submodule results in a sign-in redirect in the dep ensure stage, which is a bug on the GitLab end.  Until one of those issues is resolved, the API can't be included as a submodule.